refactor: create dir tableIndicadores
